+++ src/PersonalFinance.Application/ViewModel/Request/Expenditure/UpdateExpendidureRequest.cs
@@ -20,5 +20,5 @@ public class UpdateExpendidureRequest
     [MaxLength(60)]
     public string Description { get; set; }
 
-    public int ExpenditureCategoryId { get; set; }
+    public long ExpenditureCategoryId { get; set; }
 }
